migrate DESC TABLE AS JSON to v2 command
