+++ sql/core/src/main/scala/org/apache/spark/sql/execution/command/tables.scala
@@ -19,16 +19,12 @@ package org.apache.spark.sql.execution.command
 
 import java.net.{URI, URISyntaxException}
 
-import scala.collection.mutable
 import scala.collection.mutable.ArrayBuffer
 import scala.jdk.CollectionConverters._
 import scala.util.control.NonFatal
 
 import org.apache.hadoop.fs.{FileContext, FsConstants, Path}
 import org.apache.hadoop.fs.permission.{AclEntry, AclEntryScope, AclEntryType, FsAction, FsPermission}
-import org.json4s._
-import org.json4s.JsonAST.JObject
-import org.json4s.jackson.JsonMethods._
 
 import org.apache.spark.sql.{Row, SparkSession}
 import org.apache.spark.sql.catalyst.{SQLConfHelper, TableIdentifier}
@@ -747,261 +743,6 @@ case class DescribeTableCommand(
       case _ =>
     }
     table.storage.toLinkedHashMap.foreach(s => append(buffer, s._1, s._2, ""))
-  }
-}
-
-/**
- * Command that looks like
- * {{{
- *   DESCRIBE [EXTENDED|FORMATTED] table_name partitionSpec? [AS JSON];
- * }}}
- */
-case class DescribeTableJsonCommand(
-   table: TableIdentifier,
-   partitionSpec: TablePartitionSpec,
-   isExtended: Boolean) extends LeafRunnableCommand {
-  override val output = DescribeCommandSchema.describeJsonTableAttributes()
-  // Already added to jsonMap in DescribeTableJsonCommand
-  private val excludedKeys = Set("catalog", "schema", "database", "table")
-
-  override def run(sparkSession: SparkSession): Seq[Row] = {
-    val jsonMap = mutable.LinkedHashMap[String, JValue]()
-    val catalog = sparkSession.sessionState.catalog
-
-    if (catalog.isTempView(table)) {
-      if (partitionSpec.nonEmpty) {
-        throw QueryCompilationErrors.descPartitionNotAllowedOnTempView(table.identifier)
-      }
-      val schema = catalog.getTempViewOrPermanentTableMetadata(table).schema
-      describeColsJson(schema, jsonMap, header = false)
-    } else {
-      val metadata = catalog.getTableRawMetadata(table)
-      val schema = if (metadata.schema.isEmpty) {
-        // In older versions of Spark,
-        // the table schema can be empty and should be inferred at runtime.
-        sparkSession.table(metadata.identifier).schema
-      } else {
-        metadata.schema
-      }
-
-      addKeyValueToMap("table_name", JString(metadata.identifier.table), jsonMap)
-      table.catalog.foreach(catalog => addKeyValueToMap("catalog_name", JString(catalog), jsonMap))
-      table.database.foreach { db =>
-        addKeyValueToMap("namespace", JArray(List(JString(db))), jsonMap)
-        addKeyValueToMap("schema_name", JString(db), jsonMap)
-      }
-
-      describeColsJson(schema, jsonMap, header = false)
-      describeClusteringInfoJson(metadata, jsonMap)
-
-      if (partitionSpec.nonEmpty) {
-        // Outputs the partition-specific info for the DDL command:
-        // "DESCRIBE [EXTENDED|FORMATTED] table_name PARTITION (partitionVal*)"
-        describePartitionInfoJson(sparkSession, catalog, metadata, jsonMap)
-      } else {
-        describeFormattedTableInfoJson(metadata, jsonMap)
-      }
-    }
-
-    Seq(Row(compact(render(JObject(jsonMap.toList)))))
-  }
-
-  private def addKeyValueToMap(
-      key: String,
-      value: JValue,
-      jsonMap: mutable.LinkedHashMap[String, JValue]): Unit = {
-    // Rename some JSON keys that are pre-named in describe table implementation
-    val renames = Map(
-      "inputformat" -> "input_format",
-      "outputformat" -> "output_format"
-    )
-
-    val normalizedKey = key.toLowerCase().replace(" ", "_")
-    val renamedKey = renames.getOrElse(normalizedKey, normalizedKey)
-
-    if (!jsonMap.contains(renamedKey) && !excludedKeys.contains(renamedKey)) {
-      jsonMap += renamedKey -> value
-    }
-  }
-
-  /**
-   * Util to recursively form JSON string representation of data type, used for DESCRIBE AS JSON.
-   * Differs from `json` in DataType.scala by providing additional fields for some types.
-   */
-  private def jsonType(
-      dataType: DataType): JValue = {
-    dataType match {
-      case arrayType: ArrayType =>
-        JObject(
-          "name" -> JString("array"),
-          "element_type" -> jsonType(arrayType.elementType),
-          "element_nullable" -> JBool(arrayType.containsNull)
-        )
-
-      case mapType: MapType =>
-        JObject(
-          "name" -> JString("map"),
-          "key_type" -> jsonType(mapType.keyType),
-          "value_type" -> jsonType(mapType.valueType),
-          "value_nullable" -> JBool(mapType.valueContainsNull)
-        )
-
-      case structType: StructType =>
-        val fieldsJson = structType.fields.map { field =>
-          val baseJson = List(
-            "name" -> JString(field.name),
-            "type" -> jsonType(field.dataType),
-            "nullable" -> JBool(field.nullable)
-          )
-          val commentJson = field.getComment().map(comment => "comment" -> JString(comment)).toList
-          val defaultJson =
-            field.getCurrentDefaultValue().map(default => "default" -> JString(default)).toList
-
-          JObject(baseJson ++ commentJson ++ defaultJson: _*)
-        }.toList
-
-        JObject(
-          "name" -> JString("struct"),
-          "fields" -> JArray(fieldsJson)
-        )
-
-      case decimalType: DecimalType =>
-        JObject(
-          "name" -> JString("decimal"),
-          "precision" -> JInt(decimalType.precision),
-          "scale" -> JInt(decimalType.scale)
-        )
-
-      case varcharType: VarcharType =>
-        JObject(
-          "name" -> JString("varchar"),
-          "length" -> JInt(varcharType.length)
-        )
-
-      case charType: CharType =>
-        JObject(
-          "name" -> JString("char"),
-          "length" -> JInt(charType.length)
-        )
-
-      // Only override TimestampType; TimestampType_NTZ type is already timestamp_ntz
-      case _: TimestampType =>
-        JObject("name" -> JString("timestamp_ltz"))
-
-      case yearMonthIntervalType: YearMonthIntervalType =>
-        def getFieldName(field: Byte): String = YearMonthIntervalType.fieldToString(field)
-
-        JObject(
-          "name" -> JString("interval"),
-          "start_unit" -> JString(getFieldName(yearMonthIntervalType.startField)),
-          "end_unit" -> JString(getFieldName(yearMonthIntervalType.endField))
-        )
-
-      case dayTimeIntervalType: DayTimeIntervalType =>
-        def getFieldName(field: Byte): String = DayTimeIntervalType.fieldToString(field)
-
-        JObject(
-          "name" -> JString("interval"),
-          "start_unit" -> JString(getFieldName(dayTimeIntervalType.startField)),
-          "end_unit" -> JString(getFieldName(dayTimeIntervalType.endField))
-        )
-
-      case _ =>
-        JObject("name" -> JString(dataType.typeName))
-    }
-  }
-
-  private def describeColsJson(
-      schema: StructType,
-      jsonMap: mutable.LinkedHashMap[String, JValue],
-      header: Boolean): Unit = {
-    val columnsJson = jsonType(StructType(schema.fields))
-      .asInstanceOf[JObject].find(_.isInstanceOf[JArray]).get
-    addKeyValueToMap("columns", columnsJson, jsonMap)
-  }
-
-  private def describeClusteringInfoJson(
-      table: CatalogTable, jsonMap: mutable.LinkedHashMap[String, JValue]): Unit = {
-    table.clusterBySpec.foreach { clusterBySpec =>
-      val clusteringColumnsJson: JValue = JArray(
-        clusterBySpec.columnNames.map { fieldNames =>
-          val nestedFieldOpt = table.schema.findNestedField(fieldNames.fieldNames.toIndexedSeq)
-          assert(nestedFieldOpt.isDefined,
-            "The clustering column " +
-              s"${fieldNames.fieldNames.map(quoteIfNeeded).mkString(".")} " +
-              s"was not found in the table schema ${table.schema.catalogString}."
-          )
-          val (path, field) = nestedFieldOpt.get
-          JObject(
-            "name" -> JString((path :+ field.name).map(quoteIfNeeded).mkString(".")),
-            "type" -> jsonType(field.dataType),
-            "comment" -> field.getComment().map(JString).getOrElse(JNull)
-          )
-        }.toList
-      )
-      addKeyValueToMap("clustering_information", clusteringColumnsJson, jsonMap)
-    }
-  }
-
-  private def describeFormattedTableInfoJson(
-      table: CatalogTable, jsonMap: mutable.LinkedHashMap[String, JValue]): Unit = {
-    table.bucketSpec match {
-      case Some(spec) =>
-        spec.toJsonLinkedHashMap.foreach { case (key, value) =>
-          addKeyValueToMap(key, value, jsonMap)
-        }
-      case _ =>
-    }
-    table.storage.toJsonLinkedHashMap.foreach { case (key, value) =>
-      addKeyValueToMap(key, value, jsonMap)
-    }
-
-    val filteredTableInfo = table.toJsonLinkedHashMap
-
-    filteredTableInfo.map { case (key, value) =>
-      addKeyValueToMap(key, value, jsonMap)
-    }
-  }
-
-  private def describePartitionInfoJson(
-     spark: SparkSession,
-     catalog: SessionCatalog,
-     metadata: CatalogTable,
-     jsonMap: mutable.LinkedHashMap[String, JValue]): Unit = {
-    if (metadata.tableType == CatalogTableType.VIEW) {
-      throw QueryCompilationErrors.descPartitionNotAllowedOnView(table.identifier)
-    }
-
-    DDLUtils.verifyPartitionProviderIsHive(spark, metadata, "DESC PARTITION")
-    val normalizedPartSpec = PartitioningUtils.normalizePartitionSpec(
-      partitionSpec,
-      metadata.partitionSchema,
-      table.quotedString,
-      spark.sessionState.conf.resolver)
-    val partition = catalog.getPartition(table, normalizedPartSpec)
-
-    // First add partition details to jsonMap.
-    // `addKeyValueToMap` only adds unique keys, so this ensures the
-    // more detailed partition information is added
-    // in the case of duplicated key names (e.g. storage_information).
-    partition.toJsonLinkedHashMap.map { case (key, value) =>
-      addKeyValueToMap(key, value, jsonMap)
-    }
-
-    metadata.toJsonLinkedHashMap.map { case (key, value) =>
-      addKeyValueToMap(key, value, jsonMap)
-    }
-
-    metadata.bucketSpec match {
-      case Some(spec) =>
-        spec.toJsonLinkedHashMap.map { case (key, value) =>
-          addKeyValueToMap(key, value, jsonMap)
-        }
-      case _ =>
-    }
-    metadata.storage.toJsonLinkedHashMap.map { case (key, value) =>
-      addKeyValueToMap(key, value, jsonMap)
-    }
   }
 }
