+++ sql/core/src/main/scala/org/apache/spark/sql/execution/command/views.scala
@@ -793,7 +793,6 @@ object ViewHelper extends SQLConfHelper with Logging {
       originalText: String,
       tempFunctions: Seq[String]): CatalogTable = {
 
-    val catalog = session.sessionState.catalog
     val tempViews = collectTemporaryViews(analyzedPlan)
     val tempVariables = collectTemporaryVariables(analyzedPlan)
     // TBLPROPERTIES is not allowed for temporary view, so we don't use it for
@@ -807,7 +806,9 @@ object ViewHelper extends SQLConfHelper with Logging {
       tableType = CatalogTableType.VIEW,
       storage = CatalogStorageFormat.empty,
       schema = viewSchema,
+      viewOriginalText = Some(originalText),
       viewText = Some(originalText),
+      createVersion = org.apache.spark.SPARK_VERSION,
       properties = newProperties)
   }
